Add new file for seperation of grid levels
